pyproject.toml
[project]
name = "ft-datastar"
version = "0.1.0"
description = "DataStar integration for FastHTML"
readme = "README.md"
authors = [
    { name = "firefly", email = "promptsiren@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "datastar-py>=0.4.2",
    "python-fasthtml>=0.12.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ft_datastar"]